Refactor MySQLNodeInstance
